rename(patterns): Shells
