+++ src/patterns/shells/index.js
@@ -1,5 +1,5 @@
 export { default as Shell1 } from './shell1';
 export { default as Shell2 } from './shell2';
+export { default as Shell3 } from './shell3';
 export { default as Shell4 } from './shell4';
 export { default as Shell5 } from './shell5';
-export { default as Shell6 } from './shell6';
