+++ src/patterns/shells/shell5.js
@@ -18,7 +18,7 @@ import {
 import { home, cog, plugins, key, external, trendingUp } from "@wordpress/icons";
 
 /**
- * Render Subscription Shell
+ * Render Shell 5
  */
 function Shell5() {
     const [component, setComponent] = useState('Dashboard'); // replace string with your component do it with all buttons too this will update component on click
@@ -27,42 +27,37 @@ function Shell5() {
     };
     return (
         <>
-            <Card style={{ backgroundColor: "black", padding: "5px" }}>
+            <Card style={{ backgroundColor: "#7b7b7b4f", padding: "5px" }}>
                 <HStack>
                     <CardBody style={{paddingBottom:'200px', width:"30%" }}>
                         <VStack spacing={4}>
-                            <Heading style={{color:'white',marginLeft:'10px'}}>WPUI</Heading>
+                            <Heading style={{marginLeft:'10px'}}>WPUI</Heading>
                             <VStack expanded={false} spacing={2}>
                                 <Button 
-                                    style={{ color: "white" }}
                                     icon={home}
                                     onClick={() => handleButtonClick('Dashboard')}
                                 >
                                     Dashboard
                                 </Button>
                                 <Button 
-                                    style={{ color: "white" }}
                                     icon={cog}
                                     onClick={() => handleButtonClick('Settings')}
                                 >
                                     Settings
                                 </Button>
                                 <Button 
-                                    style={{ color: "white" }}
                                     icon={plugins}
                                     onClick={() => handleButtonClick('Addons')}
                                 >
                                     Addons
                                 </Button>
                                 <Button 
-                                    style={{ color: "white" }}
                                     icon={key}
                                     onClick={() => handleButtonClick('License')}
                                 >
                                     License
                                 </Button>
                                 <Button 
-                                    style={{ color: "white" }}
                                     icon={external}
                                     onClick={() => handleButtonClick('Support')}
                                 >
@@ -71,8 +66,8 @@ function Shell5() {
                             </VStack>
                         </VStack>
                     </CardBody>
-                    <CardBody size="large" style={{ height: "500px", width: "100%", backgroundColor:'white' }}>
-                        <HStack className='card'>
+                    <CardBody size="large" style={{ borderRadius:'10px', height: "500px", width: "100%", backgroundColor:'white' }}>
+                        <HStack>
                             <Heading>{component}</Heading>
                             {/* Display your Component Here */}
                             <HStack expanded={false}>
